Clean-up CanvasKit canvas sizing. (flutter/engine#20387)
* Clean-up CanvasKit canvas sizing.

* cache size correctly

+++ engine/src/flutter/lib/web_ui/lib/src/engine/compositor/canvaskit_api.dart
@@ -76,8 +76,8 @@ class CanvasKit {
   external SkGrContext MakeGrContext(int glContext);
   external SkSurface MakeOnScreenGLSurface(
     SkGrContext grContext,
-    double width,
-    double height,
+    int width,
+    int height,
     SkColorSpace colorSpace,
   );
   external SkSurface MakeSWCanvasSurface(html.CanvasElement canvas);
